Moved firebase get cert
