+++ src/main/kotlin/app/revanced/patches/grindr/patch/FirebaseGetCertPatch.kt
